+++ packages/frontend/src/pages/QuantifyPeriodReceiver/components/DismissDialog.tsx
@@ -1,8 +1,12 @@
-import { faCalculator, faTimes } from '@fortawesome/free-solid-svg-icons';
+import {
+  faCalculator,
+  faMinusCircle,
+  faTimes,
+} from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import ScrollableDialog from '@/components/ScrollableDialog';
 import { PraiseDto } from 'api/dist/praise/types';
-import MarkDismissedButton from './MarkDismissedButton';
+import IconButton from '@/components/IconButton';
 
 const getPraisesString = (praises: PraiseDto[]): string =>
   praises.map((p) => `#${p._id.slice(-5)}`).join(', ');
@@ -40,7 +44,9 @@ const PoolDismissDialog = ({
           </p>
           <p className="text-center">{getPraisesString(praises)}</p>
           <div className="flex justify-center">
-            <MarkDismissedButton
+            <IconButton
+              icon={faMinusCircle}
+              text="Dismiss"
               onClick={(): void => {
                 onConfirm();
                 onClose();